Build-up the 'vim-syntax' directory structure. Add support for C and C++ functions.
Only way to make Vim take into account user-defined syntax is to
follow the hard-coded rules (i.e. syntax file must match the filetype or at least
directory needs to be named after that).

Important thing about function support is not to clutter the scope
with private or protected functions (C++). Destructors are also not
very interesting in this use-case. Probably there will be more
candidates to filter out, like operators and alike.