Initial implementation of new syntax highlight strategy.
Previous idea was to provide a generic way to generate the Vim syntax files for a given source files.
That was mainly inspired by the fact to enable the support for syntax highlighting of common 3rd party
libraries such as Boost as it would be really tedious to maintain these files by hand. Therefore,
Vim syntax file generator has been introduced. Idea was to run the generator during the installation
of whole environment.

However, soon it became obvious that the same mechanism could be easily extended to enable syntax
highlighting *on-the-fly* for *any* source file given. Specifically, this means that this technique
can be applied to generate the Vim syntax files even for the code currently being developed.

This commit brings in the default implementation of such on-the-fly strategy.